+++ variants/backend-base/config/application.rb
