+++ variants/backend-base/config/database.yml.tt
