+++ variants/backend-base/config/initializers/check_env.rb
