+++ variants/backend-base/config/initializers/content_security_policy.rb
