+++ variants/backend-base/config/initializers/generators.rb
