+++ variants/backend-base/config/initializers/health_checks.rb
