+++ variants/backend-base/config/initializers/sentry.rb
