+++ variants/backend-base/config/initializers/version.rb
